Process wide current_call, convert logging in all threads into events.
In order to emit events, functions use the emit() method of a Call. The
helper async.current_call() returns the Call object that transfers emitted
events back to the creator of the Call. Previously this was always a thread
local, but it's convenient for multiple threads to be able to get access to
the Call (in particular when logging happens in multiple threads) so now it
can be set to process wide.

Generally there's a process wide 'current_call' when the commandline tools
executes, which captures all logging and turns it into events. For the
toolkit, there's a process wide 'current_call' per build process.